Trigger rating calc after new data updated. some bug fix

+++ mgxhub/rating/__main__.py
@@ -4,20 +4,26 @@ import os
 import sys
 import time
 import argparse
-from datetime import datetime
 from sqlalchemy import create_engine
 from sqlalchemy.orm import Session
 from mgxhub.config import cfg
 from mgxhub.logger import logger
 from . import EloCalculator
 
-if __name__ == "__main__":
+
+def main(db_path: str, duration_threshold: str, batch_size: str):
+    '''Main function for the standalone executable of elo rating calculator.'''
+
     # Check if the lock file exists, other processes can read the RATING_CALC_LOCK_FILE environment variable to get the lock file path
     LOCK_FILE = cfg.get('rating', 'lockfile')
 
     if os.path.exists(LOCK_FILE):
         logger.debug("Only one instance of the ELO rating calculator can run at a time. Exiting.")
         sys.exit(1)
+
+    if os.path.exists(LOCK_FILE + ".scheduled"):
+        logger.debug("The ELO rating will be processed, remove the scheduled signal.")
+        os.remove(LOCK_FILE + ".scheduled")
 
     start_time = time.time()
 
@@ -33,33 +39,43 @@ if __name__ == "__main__":
         file.write(str(timestamp) + "\n")
 
     try:
-        parser = argparse.ArgumentParser(description='Update ELO ratings.')
-        parser.add_argument('--db_path', default=cfg.get('database', 'sqlite'), help='Path to SQLite database')
-        parser.add_argument('--duration_threshold', default=cfg.get('rating', 'durationthreshold'),
-                            help='Duration threshold for ELO rating update')
-        parser.add_argument('--batch_size', default=cfg.get('rating', 'batchsize'),
-                            help='Batch size for ELO rating update')
-        args = parser.parse_args()
-
-        engine = create_engine(f"sqlite:///{args.db_path}", echo=False)
+        engine = create_engine(f"sqlite:///{db_path}", echo=False)
         session = Session(engine)
         elo = EloCalculator(session)
         elo.update_ratings(
-            duration_threshold=int(args.duration_threshold),
-            batch_size=int(args.batch_size)
+            duration_threshold=int(duration_threshold),
+            batch_size=int(batch_size)
         )
         session.close()
         engine.dispose()
     finally:
-        os.remove(LOCK_FILE)
+        if os.path.exists(LOCK_FILE):
+            os.remove(LOCK_FILE)
 
     end_time = time.time()
     elapsed_time = round(end_time - start_time, 2)
-
-    # Get the current time
-    current_time = datetime.now()
 
     # Create the log message
     log_message = f"Rating calculated, duration: {elapsed_time}"
 
     logger.info(log_message)
+
+
+if __name__ == "__main__":
+    parser = argparse.ArgumentParser(description='Update ELO ratings.')
+    parser.add_argument('--db_path', default=cfg.get('database', 'sqlite'), help='Path to SQLite database')
+    parser.add_argument('--duration_threshold', default=cfg.get('rating', 'durationthreshold'),
+                        help='Duration threshold for ELO rating update')
+    parser.add_argument('--batch_size', default=cfg.get('rating', 'batchsize'),
+                        help='Batch size for ELO rating update')
+    args = parser.parse_args()
+
+    main(args.db_path, args.duration_threshold, args.batch_size)
+
+    # If another rating calculation is triggered during calculation, it will be
+    # scheduled using a scheduled lock file. When the current calculation is
+    # finished,  the scheduled calculation will be triggered and the scheduled
+    # lock file will be removed in main().
+    LOCK_FILE_SCHEDULED = cfg.get('rating', 'lockfile') + ".scheduled"
+    if os.path.exists(LOCK_FILE_SCHEDULED):
+        main(args.db_path, args.duration_threshold, args.batch_size)
